extract out the clean workspace test to its own file
Summary:
This test relies on what the root workspace is before any other code
is run. However, some of the test cases change it. If the order the
tests are run is randomized, then the test can fail if run after one
of them.

Having it on its own ensures that it always sees a pristine state.

Test Plan: Verified locally and confirmed in CI.

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
